requirements.txt
httpx[http2]>=0.27
selectolax>=0.3
feedparser>=6.0
pydantic-settings>=2.2
rich>=13.7
Jinja2>=3.1
aiofiles>=23.2
PyYAML>=6.0
